requirements.txt
fastapi==0.95.1
pydantic==1.10.7
google-cloud-firestore==2.11.0
functions-framework==3.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
uvicorn
passlib[bcrypt]